MDEV-31995 Bogus error executing PS for query using CTE with renaming of columns
After statement preparation, the item list names associated with the select lex
for each derived table are left with the name as supplied, outside the select,
by the with column list element. For example

with cte (col1, col2) as (select a as c1, b as c2 from t) ...

the select lex associated with select a as c1, b as c2 from t, looks like

select a as col1, b as col2 from t

When there are references to columns c1 or c2 in this subselect (e.g. in an
order by clause) this means that during the second execution, looking for c1
in the item list will now fail, leading to errors detailed above.

The solution is to take a copy of the original item names (after any wildcard
expansion) and reset the names back in mysql_derived_reinit().